+++ src/onegov.org/onegov/org/path.py
@@ -159,13 +159,14 @@ def get_ticket(app, handler_code, id):
 
 @OrgApp.path(model=TicketCollection, path='/tickets/{handler}/{state}')
 def get_tickets(app, handler='ALL', state='open', page=0, group=None,
-                extra_parameters=None):
+                owner=None, extra_parameters=None):
     return TicketCollection(
         app.session(),
         handler=handler,
         state=state,
         page=page,
         group=group,
+        owner=owner or '*',
         extra_parameters=extra_parameters
     )